move scale out of simple
